move psvelo to geodesy
