+++ src/geodesy/utilvelo.h
@@ -1,0 +1,21 @@
+/*
+ *    Copyright (c) 1996-2012 by G. Patau
+ *    Copyright (c) 2013-2019 by the GMT project
+ *    Donated to the GMT project by G. Patau upon her retirement from IGPG
+ *    Distributed under the Lesser GNU Public Licence
+ *    See README file for copying and redistribution conditions.
+ */
+
+/*!
+ * \file utilvelo.h
+ */
+
+#define EPSIL 0.0001
+
+void velo_get_trans (struct GMT_CTRL *GMT, double slon, double slat, double *t11, double *t12, double *t21, double *t22);
+void velo_axis2xy(double x0, double y0, double size, double pp, double dp, double pt, double dt, double *xp, double *yp, double *xt, double *yt);
+void velo_trace_arrow (struct GMT_CTRL *GMT, double slon, double slat, double dxp, double dyp, double scale, double *x1, double *y1, double *x2, double *y2);
+void velo_ellipse_convert (double sigx, double sigy, double rho, double conrad, double *eigen1, double *eigen2, double *ang) ;
+void velo_paint_ellipse (struct GMT_CTRL *GMT, double x0, double y0, double angle, double major, double minor, double scale, double t11, double t12, double t21, double t22, int polygon, struct GMT_FILL *fill, int outline);
+int velo_trace_cross (struct GMT_CTRL *GMT, double slon, double slat, double eps1, double eps2, double theta, double sscale, double v_width, double h_length, double h_width, double vector_shape, int outline, struct GMT_PEN *pen);
+void velo_paint_wedge (struct PSL_CTRL *PSL, double x0, double y0, double spin, double spinsig, double sscale, double wedge_amp, double t11, double t12, double t21, double t22, int polygon, double *rgb, int epolygon, double *ergb, int outline);
