+++ src/gmt_supplements_module.c
@@ -43,13 +43,13 @@ struct Gmt_moduleinfo {
 static struct Gmt_moduleinfo g_supplements_module[] = {
 	{"earthtide", "geodesy", "Compute grids or time-series of solid Earth tides", ">D},GG),>DL,>DS"},
 	{"gpsgridder", "geodesy", "Interpolate GPS velocity vectors using Green's functions for a thin elastic sheet", "<D{,ND(,TG(,CD)=f,GG}"},
+	{"psvelo", "geodesy", "Plot velocity vectors, crosses, and wedges on maps", "<D{,>X}"},
 	{"gshhg", "gshhg", "Extract data tables from binary GSHHS or WDBII data files", ">D}"},
 	{"img2grd", "img", "Extract a subset from an img file in Mercator or Geographic format", "<G{,GG}"},
 	{"pscoupe", "meca", "Plot cross-sections of focal mechanisms", "<D{,>X}"},
 	{"psmeca", "meca", "Plot focal mechanisms on maps", "<D{,>X}"},
 	{"pspolar", "meca", "Plot polarities on the inferior focal half-sphere on maps", "<D{,>X}"},
 	{"pssac", "meca", "Plot seismograms in SAC format on maps", ">X},RG-"},
-	{"psvelo", "meca", "Plot velocity vectors, crosses, and wedges on maps", "<D{,>X}"},
 	{"mgd77convert", "mgd77", "Convert MGD77 data to other file formats", ""},
 	{"mgd77header", "mgd77", "Create MGD77 headers from A77 files", ""},
 	{"mgd77info", "mgd77", "Extract information about MGD77 files", ""},
